Restore individual mocked methods instead of everything
When mocking an 'aws-sdk' method, 'aws-sdk-mock' would first create a
mock for the service itself and then one for the method.
When restoring everything, 'aws-sdk-mock' would first restore all
mocked methods of a service and then the service mock.

For some reason it is not possible to mock the service a second time and
by extension mock a method from it.

Calling 'restore' with both service name and method name restores only
the mocked method and not the service. This allows the service mock to
be reused and by extension another method to be mocked.